% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframewidget.R
\name{iframewidget}
\alias{iframewidget}
\title{Create an iframe widget}
\usage{
iframewidget(src, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{src}{URL to render in an iframe}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{optional ID to be assigned to element enclosing the iframe}
}
\description{
Take a specified URL and render it in an iframe
}
\examples{
iframewidget("http://purdue.edu")
}

